+++ package.json
@@ -20,16 +20,11 @@
     "encoding": "^0.1.13",
     "eslint": "8.36.0",
     "eslint-config-next": "13.2.4",
-    "export-to-csv": "^1.2.0",
-    "file-saver": "^2.0.5",
     "material-react-table": "^1.9.1",
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-export-table-to-excel": "^1.0.6",
-    "react-select": "^5.7.5",
     "tailwindcss": "^3.2.7",
-    "typescript": "5.0.2",
-    "xlsx": "^0.18.5"
+    "typescript": "5.0.2"
   }
 }
