+++ styles/defaultpage.css
@@ -1,10 +1,4 @@
 /* This is why the default background color for everything is black */
-
-:root {
-    --primary-color: #681296;
-    --secondary-color: #9654ba;
-  }
-  
 body {
     background-color: #000;
 }
