+++ styles/globals.css
@@ -2,24 +2,3 @@
 @tailwind components;
 @tailwind utilities;
 @tailwind variants;
-
-/* width */
-::-webkit-scrollbar { 
-    width: 5px; 
-} 
-  
-/* Track */
-::-webkit-scrollbar-track { 
-    background: transparent; 
-} 
-  
-/* Handle */
-::-webkit-scrollbar-thumb { 
-    background: #888; 
-    border-radius: 5px; 
-} 
-  
-/* Handle on hover */
-::-webkit-scrollbar-thumb:hover { 
-    background: #555; 
-}
