+++ styles/navbar.css
@@ -1,7 +1,3 @@
-.nav-bar{
-    z-index: 999;
-}
-
 .hover-underline-animation {
     display: inline-block;
     position: relative;
