--- styles/selectlabel.css
@@ -1,10 +1,0 @@
-label {
-    color: rgba(0, 0, 0, 0.6);
-    font-family: "Roboto","Helvetica","Arial",sans-serif;
-    font-weight: 400;
-    font-size: 0.75rem;
-    line-height: 1.4375em;
-    letter-spacing: 0.00938em;
-    padding: 0;
-    margin: 0;
-}
